src/requirements.txt
django==3.1.2
djangorestframework==3.12.1
markdown==3.3.1
django-filter==2.4.0
django-rest-auth==0.9.5
django-allauth==0.42.0
djangorestframework-jwt==1.11.0
drf-yasg2==1.18.5
packaging==20.4
django-cors-headers==3.5.0
celery==5.0.1
requests==2.24.0
psycopg2-binary==2.8.6
